Working Attempt
Initially I had tried to create this simulation with different methods for each function. Doing this helped me understand the problem more and allowed me to create the simulation within the main method.

The user is shown the same message over and over with the current drive-way order being displayed each time.

First, the reg is checked to see if it is 0. This will end the while loop ending the simulation.

If that is false the program then checks if the reg is > 0, meaning the user wants to add a car. If a reg is entered that is already in the car park the guest is removed.

Finally the program checks if the reg is < 0 meaning the user wants to retrieve their car. After the reg is set to positive, the stack is checked to see if the reg is at the top (last in). If this is false cars that are in the way are moved to the steet. Once the desired car is removed the cars on the street are placed back into driveway.