Updates to the way both stacks are displayed
Added 2 methods, printDriveway() & printStreet(), these methods are called when cars need to be moved from street to driveway and vice versa. This shows the user which order the cars would have to be moved to get the desired one.

Both methods use the Iterator class to iterate through each value with the stacks to get the reg of the car.

Message displays when 0 is entered to show the program is ending.